fix reactive components bugs and add post CRUD functionallity

+++ src/stores/Feed.svelte.ts
@@ -7,8 +7,10 @@ import type { QueryBuilder } from "./QueryBuilders/QueryBuilder";
 
 export class Feed{
     private _posts = $state<Post[]>([]);
-    private _isLoading = $state(true);
+    private _isLoading = $state(false);
     private queryBuilder : QueryBuilder;
+    private _size = $state(0);
+    private _fetchedAll = false;
 
     private static instance : Feed | null = null;
 
@@ -22,6 +24,8 @@ export class Feed{
                     this._posts.push(new Post(post.ref, post.data() as PostSchema, post.id));
                     DataBasaConn.getDB().getUserInfo(post.data().createdBy);
                 })
+                this._size += result.size;
+                this._fetchedAll = this._size < 10 ? true : false;
                 this._isLoading = false;
             });
         })
@@ -32,14 +36,21 @@ export class Feed{
         return this._isLoading;
     }
 
-
     get posts(): Post[] {
         return this._posts;
     }
 
-    async loadNewPosts() : Promise<boolean>{
+    get size() : number{
+        return this._size;
+    }
 
-        this._isLoading = true;
+    get fetchedAll() : boolean{
+        return this._fetchedAll;
+    }
+
+    async loadNewPosts() : Promise<boolean>{
+        if(this._fetchedAll) return true;
+
         const result = await getDocs(await this.queryBuilder.getFetchQuery(this._posts[this._posts.length - 1]));
 
         result.forEach(post => {
@@ -47,23 +58,23 @@ export class Feed{
             console.log(post.id)
         })
 
-        this._isLoading = false;
-        if(result.size < 10) return true;
-        else return false;
+        this._size += result.size;
+        this._fetchedAll = result.size < 10 ? true : false;
+
+        return this._fetchedAll;
     }
 
     addPost(newPost: Post) {
         this._posts = [ newPost, ...this._posts]
     }
 
-    deletePost(postId: string): void {
+    deletePost(postId: string, deleteFromDB : boolean): void {
         this._posts = this._posts.filter((post, i, posts) => {
             if (post.postId == postId && post.data.createdBy == MyUser.getUser()!.userInfo!.Username){
-                post.delete();
+                if (deleteFromDB) post.delete();
                 return false;
-            }else true;
+            }else return true;
         });
-
     }
 
 }
